requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.4
daphne==4.1.0
distro==1.9.0
Django==5.0.2
django-cors-headers==4.2.0
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.16.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.7
openai==1.12.0
phonenumbers==8.13.30
pillow==10.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
zope.interface==6.2
